requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
aiosqlite==0.20.0
sqlalchemy[asyncio]==2.0.35
python-telegram-bot==21.6
google-generativeai==0.8.0
rembg==2.0.57
Pillow==10.4.0
onnxruntime==1.19.0
pydantic==2.9.0
httpx==0.27.0
python-multipart==0.0.12
